Add isBBoxInPolygon function and update selection logic to use it
